SBERDOMA-1028 extracted soft-delete logic into `DeleteButtonWithConfirmModal` component
