+++ apps/condo/lang/en.json
@@ -474,9 +474,11 @@
   "division.form.hint.responsible.description": "Responsible for all types of works in selected properties",
   "division.form.hint.executors.title": "Executors",
   "division.form.hint.executors.description": "Do work «by hands» in selected properties",
-  "division.form.actions.delete": "Delete division",
   "division.validation.responsible.required": "Select employee, that will be responsible for this division",
   "division.validation.properties.required": "Select properties for this division",
+  "division.action.delete.confirm.title": "Are you sure to delete division?",
+  "division.action.delete.confirm.message": "After deleting the division the system will be unable to automatically assign responsible and executor for next tickets, being created on specified properties in division. Employees and properties will not be deleted.",
+  "division.action.delete.confirm.ok": "Delete division",
   "ticket.EmptyList.header": "The list of applications is still empty",
   "ticket.EmptyList.title": "Create the first one, it's simple",
   "ticket.status.OPEN.name": "Open",
