+++ apps/condo/lang/ru.json
@@ -474,9 +474,11 @@
   "division.form.hint.responsible.description": "Отвечает «головой» за все типы работ на выбранных домах",
   "division.form.hint.executors.title": "Исполнители",
   "division.form.hint.executors.description": "Выполняют «руками» работы на выбранных домах",
-  "division.form.actions.delete": "Удалить участок",
   "division.validation.responsible.required": "Выберите сотрудника, который будет ответственным за данный участок",
   "division.validation.properties.required": "Выберите дома, которые будут входить в данный участок",
+  "division.action.delete.confirm.title": "Точно удалить участок?",
+  "division.action.delete.confirm.message": "После удаления участка пропадёт возможность автоматического назначения ответственных и исполнителей для создаваемых заявок на указанные в участке дома. Сотрудники и дома удалены не будут.",
+  "division.action.delete.confirm.ok": "Удалить участок",
   "ticket.EmptyList.header": "Список заявок пока пуст",
   "ticket.EmptyList.title": "Создайте первую, это просто",
   "ticket.status.OPEN.name": "Открыта",
